unify STORAGE_TYPE and refactor storage connections
