+++ app/airflow/dags/libs/SR_processing/core.py
@@ -1,5 +1,3 @@
-from airflow.providers.microsoft.azure.hooks.wasb import WasbHook
-from airflow.providers.amazon.aws.hooks.s3 import S3Hook
 from pathlib import Path
 from typing import Dict, Any
 import logging
@@ -9,72 +7,14 @@ from openpyxl.worksheet.worksheet import Worksheet
 from libs.utils import (
     pull_validated_params,
 )
-from airflow.models.connection import Connection
-from airflow.utils.session import create_session
+
 import os
-
-logger = logging.getLogger(__name__)
-storage_type = os.getenv("AIRFLOW_VAR_STORAGE_TYPE", "minio").lower()
-
-
-def connect_to_storage():
-
-    with create_session() as session:
-        if storage_type == "azure":
-            # Check if WASB connection exists
-            existing_conn = (
-                session.query(Connection)
-                .filter(Connection.conn_id == "wasb_conn")
-                .first()
-            )
-
-            if existing_conn is None:
-                conn = Connection(
-                    conn_id="wasb_conn",
-                    conn_type="wasb",
-                    extra={
-                        "connection_string": os.getenv(
-                            "AIRFLOW_VAR_WASB_CONNECTION_STRING"
-                        ),
-                    },
-                )
-                session.add(conn)
-                session.commit()
-                logger.info("Created new WASB connection")
-
-        elif storage_type == "minio":
-            # Check if MinIO connection exists
-            existing_conn = (
-                session.query(Connection)
-                .filter(Connection.conn_id == "minio_conn")
-                .first()
-            )
-
-            if existing_conn is None:
-                conn = Connection(
-                    conn_id="minio_conn",
-                    conn_type="aws",
-                    extra={
-                        "endpoint_url": os.getenv("AIRFLOW_VAR_MINIO_ENDPOINT"),
-                        "aws_access_key_id": os.getenv("AIRFLOW_VAR_MINIO_ACCESS_KEY"),
-                        "aws_secret_access_key": os.getenv(
-                            "AIRFLOW_VAR_MINIO_SECRET_KEY"
-                        ),
-                    },
-                )
-                session.add(conn)
-                session.commit()
-                logger.info("Created new MinIO connection")
+from libs.enums import StorageType
+from libs.utils import get_storage_hook
 
 
-def _get_storage_hook():
-
-    if storage_type == "azure":
-        return WasbHook(wasb_conn_id="wasb_conn")
-    elif storage_type == "minio":
-        return S3Hook(aws_conn_id="minio_conn")
-    else:
-        raise ValueError(f"Unsupported storage type: {storage_type}")
+storage_type = os.getenv("STORAGE_TYPE", StorageType.MINIO)
+hook = get_storage_hook()
 
 
 def process_scan_report_task(**kwargs):
@@ -92,25 +32,24 @@ def process_scan_report_task(**kwargs):
     validated_params = pull_validated_params(kwargs, "validate_params_SR_processing")
     scan_report_blob = validated_params["scan_report_blob"]
     local_file_path = Path(f"/tmp/{scan_report_blob}")
-    hook = _get_storage_hook()
 
     try:
         # Download file based on storage type
-        logger.info(f"Downloading file from {container_name}/{scan_report_blob}")
+        logging.info(f"Downloading file from {container_name}/{scan_report_blob}")
 
-        if storage_type == "azure":
+        if storage_type == StorageType.AZURE:
             hook.get_file(
                 file_path=local_file_path,
                 container_name=container_name,
                 blob_name=scan_report_blob,
             )
-        else:  # MinIO/S3
+        elif storage_type == StorageType.MINIO:
             s3_object = hook.get_key(key=scan_report_blob, bucket_name=container_name)
             with open(local_file_path, "wb") as f:
                 s3_object.download_fileobj(f)
 
         # Read and process the Excel file
-        logger.info(f"Reading file from {local_file_path}")
+        logging.info(f"Reading file from {local_file_path}")
         workbook = load_workbook(filename=local_file_path, read_only=True)
         worksheet = workbook.active
 
@@ -132,7 +71,7 @@ def process_scan_report_task(**kwargs):
         return processed_data
 
     except Exception as e:
-        logger.error(f"Error processing scan report: {str(e)}")
+        logging.error(f"Error processing scan report: {str(e)}")
         raise
     finally:
         # Clean up
