+++ app/airflow/dags/libs/enums.py
@@ -1,0 +1,6 @@
+from enum import StrEnum
+
+
+class StorageType(StrEnum):
+    AZURE = "azure"
+    MINIO = "minio"
